[optimisation]: optimised basic features like explain featues, run search query, tell time, wikipedia search.
